+++ src/components/Profile/saved-recipe-item.js
@@ -2,14 +2,17 @@ import { ListItem, Flex, Image, Text, IconButton } from "@chakra-ui/react";
 import { FaHeart } from "react-icons/fa";
 import {useDispatch} from "react-redux";
 import {removeFavoriteThunk} from "../../services/favorite-thunks";
-import {Link} from "react-router-dom";
+import {Link, useParams} from "react-router-dom";
 import React from "react";
 
 const SavedRecipeItem = ({favorite}) => {
+    const {uid} = useParams();
     const dispatch = useDispatch();
     const removeFavoriteHandler = (favorite) => {
-        console.log('removeFavoriteHandler', favorite)
-        dispatch(removeFavoriteThunk(favorite))
+        if (!uid) {
+            console.log('removeFavoriteHandler', favorite)
+            dispatch(removeFavoriteThunk(favorite))
+        }
     }
     return (
         <ListItem mt={3}>
